Move the argument parser generation into a re-usable module.
